Fix the snprintf warnings for everyone this time
== DETAILS
So, basically this back-and-forth is because we used fixed-size
data types (i.e. `uint32_t`) which maps to different primitive data
types on different platforms. So `uint32_t` might be a `long` on some
platforms (e.g. Wii U), while it's just a plain integer on others (PC).
And the format specifier works off primitive data type, not data type
size.

So, to resolve this, we:

- keep `%lx` as the format specifier
- cast the parameter to printf to unsigned long

This is better than the alternatives that could cause problems trying to
cast a long down to an int.

+++ core_backup.c
@@ -160,7 +160,7 @@ bool core_backup_get_backup_path(
          (unsigned)time_info.tm_hour,
          (unsigned)time_info.tm_min,
          (unsigned)time_info.tm_sec,
-         crc,
+         (unsigned long)crc,
          (unsigned)backup_mode,
          FILE_PATH_CORE_BACKUP_EXTENSION);
 
@@ -748,6 +748,6 @@ bool core_backup_list_get_entry_crc_str(
    if (!entry || (len < 9))
       return false;
 
-   snprintf(crc, len, "%08lx", entry->crc);
+   snprintf(crc, len, "%08lx", (unsigned long)entry->crc);
    return true;
 }
